requirements.txt
fastapi==0.133.0
uvicorn==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
alembic==1.18.4
python-dotenv==1.2.2
pydantic-settings==2.6.1
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
